backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
pyttsx3==2.90
SpeechRecognition==3.10.0
google-generativeai==0.3.2
mtranslate==1.8.2
python-socketio==5.9.0
eventlet==0.33.3 